Fase1/_backup_phase1_base/phase1_auth-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
asyncpg==0.30.0
pydantic-settings==2.8.1
PyJWT==2.10.1
bcrypt==4.3.0
email-validator==2.2.0
python-dotenv==1.0.1

pytest==8.3.4
httpx==0.28.1
